Grand Exchange Progress
- Added OSRS items to search.
- Filtered out untradables and duplicates.
- Converted prices to long.
- Added proper price formatting.
- Added 't' key handling for trillions in enter amount input.

+++ src/com/simplicity/client/widget/listener/WidgetEnterInputListener.java
@@ -2,6 +2,6 @@ package com.simplicity.client.widget.listener;
 
 public interface WidgetEnterInputListener {
 	
-	void onEnterAmount(int widgetId, int amount);
+	void onEnterAmount(int widgetId, long amount);
 	
 }
